# [i2c] Split the target and controller FSMs