--- a/hw/ip/i2c/i2c.core
+++ b/hw/ip/i2c/i2c.core
@@ -19,7 +19,8 @@ filesets:
       - rtl/i2c_fifo_sync_sram_adapter.sv
       - rtl/i2c_fifos.sv
       - rtl/i2c_core.sv
-      - rtl/i2c_fsm.sv
+      - rtl/i2c_controller_fsm.sv
+      - rtl/i2c_target_fsm.sv
       - rtl/i2c.sv
     file_type: systemVerilogSource