… hw/ip/i2c/rtl/i2c_controller_fsm.sv
… hw/ip/i2c/rtl/i2c_target_fsm.sv