@a-will 2024-03-29
At least for now, this is just a draft to show how it could be done.