@a-will 2024-04-03
Added a commit to implement the controller halt events CSR as described before. It also at least attempts to eliminate some race conditions with software modifying CSRs during automatic Stop generation.

I did switch back to having SCL held low. If the Stop is going to happen automatically anyway, this seems acceptable, since it puts SCL in the correct place to prepare for a Stop (or Start, if software handles it and prefers).