@a-will 2024-04-05
For a gauge of the area impact, here are the FPGA results:

| Resource | Before | After |
| - | - | - |
| Flops | 1029 | 1102 |
| LUTs | 1612 | 1588 | 

The LUTs decrease as the FSMs get a little simpler, and the flops increase. Most likely, the flop increase is largely due to the duplication of 2 counters:

- 20-bit `tcount_q`
    - Used to count out hold times on driven bits and setup times for stretching on the target.
    - Used extensively on the host.
    - This counter is larger than necessary: It only need to be capable of a max count of 5 us.
        -  Even if we support up to a 1 GHz core clock, 13 bits would be enough.
- 32-bit `stretch_idle_cnt`
    - For the target, used to count a timeout for bus idle
    - For the host, used to count how long a target has stretched the clock
    - This counter is needlessly large: It only need to be capable of counting to 35 ms.
        - Even if we support up to a 1 GHz core clock, 26 bits would be enough.