@a-will 2024-04-05
The DV failure is unrelated (in LC_CTRL block level in the VOLATILE_UNLOCK_RAW configuration).