@a-will 2024-04-06
> Code changes LGTM, glad this task ended up being not too difficult, as there wasn't too much shared state after all. I only have one meta question before merging. Do we want to merge these changes into master? I understood there was some discussion about derisking by keeping the M2 design for some instances, and then working on all extensions as a separate block. In that case, maybe we want to target these changes into a separate branch, or maybe even a new copy of the block itself? @andreaskurth @mundaym may have some input here.

I don't think that strategy makes sense, since it suggests the current i2c block is acceptable in any way. There isn't some proven IP to fall back on here. Make no mistake: The current code has numerous bugs that I have started fixing, and the regression results reveal an IP that just is not at tape-out quality.

Trying to have two separate IPs only adds overhead in a time when we can't afford it. We don't have the resources.

Instead, I am cleaning up the IP in ways that make it more robust for single controller environments. In the series of PRs that follow, there are fixes to existing features we targeted for D2 sign-off, and the DV tests substantially improve. Splitting the FSM into two simplifies the code and makes it easier to work with, and it enables us to share fixes.

Meanwhile, it looks like area will not be significantly impacted, as the following cleanup PRs reduce it from here (have to check the area usage again after the last commit for the target mode cleanups, though).

In any case, please don't hold up my PRs for that discussion if it needs to continue. The older version can be retrieved from the git history if necessary.