@andreaskurth 2024-04-08
Thanks @a-will for this work and @hcallahan-lowrisc for reviewing. :+1:

These changes can go into the i2c HW IP block on `master`.